Split Media.swift into multiple files to make code management easier
